app: more versatile initialization for path in config
in case a manifest file is contained into some sub directories,
the command to init a workspace locally should be
```
west init -l sub1/sub2 --mf west.yml
```

but in that case, the west top dir is going to be into `sub1`
which is an issue.

one workaround is to use:
```
west init -l sub1 --mf sub2/west.yml
```

but in that case, the config is initialized incorrectly:
`manifest.path` contains only `sub1` and `manifest.file` is `sub2/west.yml`.

`manifest.path` should contain a path, even if `--mf` is a relative path.
`manifest.file` should only contain the filename.

this change allows any usage so that the workspace top dir is created next to `sub1`
and the manifest repo can be located into nested directories

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>