tests: Extend test_extension_command_duplicate
This test is meant to cover the handling of duplicate command names,
including the warnings to be emitted. Previously, it only tested the
case where two extension commands had the same name. The other case is
duplicating a built-in command name, which prompts a different warning
message, so update the test to include it.

Additionally, ensure that each duplicate command in the test comes from
a different project, so that each warning message is expected to contain
a different project name. This serves to verify the bug fix from commit
edd1cee2ebb6191cb6fa4ec874d0113c1fd2b9a3.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>