repo: Add strict to zip calls
By default, if iterables passed to zip have different lengths, the
resulting iterator will be silently truncated to the length of the
shortest.
The `strict` argument was introduced in Python 3.10 and should be explicit
for all zip function calls. This is also verified with the
flake8-bugbear linter checks in ruff.

Signed-off-by: Pieter De Gendt <pieter.degendt@basalte.be>