test_project.py: fix test not to require a 500MB download.
Fix "test_init_with_clone_option_depth_one(repos_tmpdir)" to actually use
the local repos_tmpdir fixture instead of requiring a network connection
to github.com/zephyr and using it to download rougly 500 MB (as of today).

This saves considerable time on slow connnections and makes it possible
to run the whole test suite offline (when all dependencies are already
available).

Fixes commit 1007557b92cad ("Add test cases for west init -o= option")

This wasn't obvious and was missed in code review because
github.com/zephyr is the default manifest for "west init".

(and yes this was found on a plane...)

Signed-off-by: Marc Herbert <marc.herbert@intel.com>