support config dropin directories
config files in acccording dropin directories are considered
automatically. The dropin directory is named as the config itself but
with a '.d' suffix. For example for local config the dropin directory is
'./west/config.d' and for global config it is '~/.westconfig.d'

support config dropin directories

config files in acccording dropin directories are considered
automatically. The dropin directory is named as the config itself but
with a '.d' suffix. For example for local config the dropin directory is
'./west/config.d' and for global config it is '~/.westconfig.d'

Boolean option `config.dropins` must be set to `true` so that dropin
configs for the according config type are considered. So if the option
config.dropins=true in enabled in local config, only local dropin
configs are used (but no system or global ones).