configuration: anchor relative WEST_CONFIG_x paths to "west topdir"
Fixes #871.

This could strictly speaking be considered as a backwards-compatible
change, but I don't think it should because: 1. who in their sane mind
would have intentionally relied on the behavior of their west commands
to change depending on the current directory!?

I think it is still very unlikely, but: 2. it's a bit more likely that
some people always run west commands from their topdir directory _and_
leveraged that to saved a little bit of typing with a relative
WEST_CONFIG_x definition(s). Or, maybe even used the same relative
$WEST_CONFIG_x value across workspaces, i.e: topdir1/$WEST_CONFIG_x and
topdir2/$WEST_CONFIG_x, etc. ?

This commit breaks 1. but preserves 2.

To be clear: this commit still does not "condone" relative WEST_CONFIG_x
values; I don't think we should document that. But it makes their behavior
_deterministic_ which is what matters.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>