support multiple west config files per config level
multiple config files can be specified in the according environment
variable for each config level (separated by 'os.pathsep', which is ';'
on Windows or ':' otherwise). The config files are applied in the same
order as they are specified, whereby values from later files override
earlier ones.

'config --list-paths' prints all considered config files which are
currently in use and exist.

existing tests are adopted to work with the internal change of the
return type of _location from 'str' to 'list[str]'