tests: don't run test_import_project_release_dir() in a random dir
Generally speaking, tests sensitive to the current directory should not
run from a random directory.

The current directory of test_import_project_release_dir() was either
the current directory or wherever the previous test left it, which could
be inside a random west workspace. This crashed the `west init -l
/tmp/pytest/...` command used by the test.

Found thanks to pytest-xdist and pytest -n 20, but there is more
"reliable" way to reproduce: move west.git instead a west workspace and
run:

   pytest -k test_import_project_release_dir

This fails with the "already initialized" error.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>