Improve test isolation
Extend and rename current autouse pytest fixture to
`setup_teardown_test_environment`.

The fixture ensures an isolated test environment.
It creates a new temporary directory which is used as working directory.
This ensures a clean start for each test and prevents tests from
affecting another one through changes to the working directory.

The fixture ensures that the user's actual configuration files are
neither used nor touched during test, as WEST_CONFIG_* env variables are
set, whereby no config files are created at these locations.

The fixture sets ZEPHYR_BASE (to avoid complaints in subcommand stderr),
but to a spurious location (so that attempts to read from inside of it
are caught here).

The fixture also ensures that any environment modifications made by a
test do not leak into subsequent tests, as the environment is restored
when the `update_env` with-block exits.